+++ packages/react/src/experimental/OneDataCollection/CollectionActions/CollectionActions.tsx
@@ -17,8 +17,10 @@ export const CollectionActions = ({
   const buttonActions = (primaryActions && [primaryActions]) || []
   const dropdownActions = secondaryActions || []
 
+  if (buttonActions.length === 0 && dropdownActions.length === 0) return null
+
   return (
-    <div className="flex flex-row-reverse gap-2">
+    <div className="flex flex-row-reverse items-center gap-2">
       {buttonActions.map((action) => (
         <Button
           size="md"
@@ -32,9 +34,11 @@ export const CollectionActions = ({
 
       {dropdownActions.length > 0 && (
         <Dropdown items={dropdownActions}>
-          <Button variant="ghost" icon={Ellipsis} label="Actions" hideLabel />
+          <Button variant="outline" icon={Ellipsis} label="Actions" hideLabel />
         </Dropdown>
       )}
+
+      <div className="mx-1 h-4 w-px bg-f1-border-secondary" />
     </div>
   )
 }